+++ src/features/club-application/model/useApplyClub.ts
@@ -3,8 +3,9 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import axios, { HttpStatusCode } from "axios";
 import { toast } from "sonner";
-import { clubMutations, clubQueries } from "@/entities/club/api/clubQueries";
-import type { ClubApplicationRequest } from "@/entities/club/model/club";
+import { clubQueries } from "@/entities/club/api/clubQueries";
+import { clubApplicationMutations } from "@/entities/club-application/api/clubApplicationQueries";
+import type { ClubApplicationRequest } from "@/entities/club-application/model/application";
 
 export function useApplyClub(clubId: number) {
   const queryClient = useQueryClient();
@@ -12,7 +13,7 @@ export function useApplyClub(clubId: number) {
 
   return useMutation({
     mutationFn: (body: ClubApplicationRequest) =>
-      clubMutations.applyClub(clubId, body),
+      clubApplicationMutations.apply(clubId, body),
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: detailQuery.queryKey });
       toast.success("동아리 신청이 완료되었습니다.");
