+++ src/features/club-form/model/useClubFormBuilder.ts
@@ -3,7 +3,7 @@
 import { useReducer, type FormEvent } from "react";
 import { useRouter } from "next/navigation";
 import { toast } from "sonner";
-import type { ClubForm, ClubFormFieldType } from "@/entities/club/model/club";
+import type { ClubForm, ClubFormFieldType } from "@/entities/club-form/model/form";
 import {
   clubFormDraftReducer,
   createClubFormDraftState,
